requirements.txt
alembic==1.11.3
aniso8601==9.0.1
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
click==8.1.7
colorama==0.4.6
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.4
flask-restplus==0.13.0
Flask-SQLAlchemy==3.0.5
flask-swagger-ui==4.11.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
Mako==1.2.4
MarkupSafe==2.1.3
psycopg2==2.9.7
python-dotenv==1.0.0
pytz==2023.3
referencing==0.30.2
rpds-py==0.9.2
six==1.16.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
Werkzeug==2.3.7
